RS types for validation of RS class attributes.

Supported types for RS classes include:

t_any

t_date, t_dates

t_int, t_ints

t_dbl, t_dbls

t_num, t_nums

t_char, t_chars

t_bool, t_bools

t_cplx, t_cplxs

t_raw, t_raws

t_factor, t_factors

t_list

t_array

t_vector

t_matrix

t_dataframe

t_hashtab

t_environment

t_pairlist

t_func

t_expr

t_call

t_sym

t_lang

t_obj

t_prim

Note that there are scalar and vector versions of some types, such as t_int and t_ints. The former allows for a single scalar integer, while the latter matches a vector of integers.

There is also the catch-all type t_any, which will match any value, and is useful for cases where you want to allow any type of value for a field, while still validating the other fields, i.e. when .validation = TRUE (the default).

Usage

t_any()

t_date()

t_dates()

t_int()

t_ints()

t_dbl()

t_dbls()

t_num()

t_nums()

t_char()

t_chars()

t_bool()

t_bools()

t_cplx()

t_cplxs()

t_raw()

t_raws()

t_factor()

t_factors()

t_list()

t_array()

t_vector()

t_matrix()

t_dataframe()

t_hashtab()

t_environment()

t_pairlist()

t_func()

t_expr()

t_call()

t_sym()

t_lang()

t_obj()

t_prim()

Value

A new type object that can be used for validation in RS classes.

Examples

t_any
#> function () 
#> .Call(wrap__ClassType__from_str, name, PACKAGE = "RS")
#> <bytecode: 0x55e95522b0f8>
#> <environment: 0x55e955228218>
#> attr(,"class")
#> [1] "TypeGenerator" "function"      "RS"           
t_date
#> function () 
#> .Call(wrap__ClassType__from_str, name, PACKAGE = "RS")
#> <bytecode: 0x55e955268538>
#> <environment: 0x55e955269488>
#> attr(,"class")
#> [1] "TypeGenerator" "function"      "RS"           
t_ints
#> function () 
#> .Call(wrap__ClassType__from_str, name, PACKAGE = "RS")
#> <bytecode: 0x55e955297b98>
#> <environment: 0x55e955298ae8>
#> attr(,"class")
#> [1] "TypeGenerator" "function"      "RS"           
t_dataframe
#> function () 
#> .Call(wrap__ClassType__from_str, name, PACKAGE = "RS")
#> <bytecode: 0x55e9552cac10>
#> <environment: 0x55e9552c5e10>
#> attr(,"class")
#> [1] "TypeGenerator" "function"      "RS"